% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{create_empty_package}
\alias{create_empty_package}
\title{Make a package from a DESCRIPTION file}
\usage{
create_empty_package(desc_path, ...)
}
\arguments{
\item{desc_path}{The path to a valid \code{DESCRIPTION} file.}

\item{...}{Arguments for \code{\link[pkgbuild:build]{pkgbuild::build()}}}
}
\value{
The path of the built package.
}
\description{
Make a package from a DESCRIPTION file
}
\details{
The package consists of a \code{DESCRIPTION} file and a \code{NAMESPACE} file.
Note that the \code{pkgbuild} package is required.
}
