% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{expect_df}
\alias{expect_df}
\title{Helper function to test dataframes}
\usage{
expect_df(df, expected_cols)
}
\arguments{
\item{df}{Dataframe to be tested.}

\item{expected_cols}{Named vector with entries of the form "\verb{<column name>} = \verb{<column type>}".
For example \code{c('Foo' = 'character')}.}
}
\value{
Invisble \code{TRUE}. This function is called for its side-effects, which is expectations
from \code{testthat}.
}
\description{
Helper function to test dataframes
}
\details{
See for example \url{https://blog.r-hub.io/2020/11/18/testthat-utility-belt}
}
