% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_package_archive_url}
\alias{get_package_archive_url}
\title{Get URL of package archive}
\usage{
get_package_archive_url(package_name)
}
\arguments{
\item{package_name}{Name of the package, .e.g. \verb{'dplyr}.}
}
\description{
Get URL of package archive
}
