% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{install_version}
\alias{install_version}
\title{Install package as they appeared in the past}
\usage{
install_version(versions, ...)
}
\arguments{
\item{versions}{An output from \code{\link[=get_version]{get_version()}}.}

\item{...}{Arguments passed on to \code{\link[=install.packages]{install.packages()}}.}
}
\value{
The same as \code{\link[=install.packages]{install.packages()}}.
}
\description{
Install package as they appeared in the past
}
