% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{make_cran}
\alias{make_cran}
\title{Make downloaded packages into valid CRAN}
\usage{
make_cran(versions)
}
\arguments{
\item{versions}{An output from \code{\link[=get_version]{get_version()}}.}
}
\value{
The root folder of the CRAN.
}
\description{
Make downloaded packages into valid CRAN
}
