% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{remove_base_packages}
\alias{remove_base_packages}
\title{Remove base packages}
\usage{
remove_base_packages(package_names)
}
\arguments{
\item{package_names}{Name of the package(s).}
}
\value{
The unique entries in \code{package_names} without base packages (if any).
}
\description{
It is common to have base R packages dependencies, but when trying to install a base package in
RStudio you'll get complaints about the package already being loaded.
This function removes the base R packages from the input (if any).
}
