% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_package_versions}
\alias{scrape_package_versions}
\title{Scrape package versions}
\usage{
scrape_package_versions(package_name)
}
\arguments{
\item{package_name}{Name of the package as a string.}
}
\value{
A dataframe with columns \code{Name} (package name and version) and \code{LastModified}.
}
\description{
Scrape package versions from CRAN.
}
